added csrf

+++ src/webapp/controllers/UserController.php
@@ -53,47 +53,54 @@ class UserController extends Controller
 
     function delete($tuserid)
     {
-        if(Auth::userAccess($tuserid))
-        {
-            $user = User::findById($tuserid);
-            $user->delete();
-            $this->app->flash('info', 'User ' . $user->getUsername() . '  with id ' . $tuserid . ' has been deleted.');
-            $this->app->redirect('/admin');
-        } else {
-            $username = Auth::user()->getUserName();
-            $this->app->flash('info', 'You do not have access this resource. You are logged in as ' . $username);
-            $this->app->redirect('/');
-        }
+        
+            if(Auth::userAccess($tuserid))
+            {   
+                $user = User::findById($tuserid);
+                $user->delete();
+                $this->app->flash('info', 'User ' . $user->getUsername() . '  with id ' . $tuserid . ' has been deleted.');
+                $this->app->redirect('/admin');
+            } else {
+                $username = Auth::user()->getUserName();
+                $this->app->flash('info', 'You do not have access this resource. You are logged in as ' . $username);
+                $this->app->redirect('/');
+            }
     }
 
     function deleteMultiple()
     {
-      if(Auth::isAdmin()){
-          $request = $this->app->request;
-          $userlist = $request->post('userlist'); 
-          $deleted = [];
+        $request = $this->app->request;
+        $usrStr = $request->post('str');
+        if($usrStr == $_SESSION["randStr"]) {
+              
+            if(Auth::isAdmin()){
 
-          if($userlist == NULL){
-              $this->app->flash('info','No user to be deleted.');
-          } else {
-               foreach( $userlist as $duserid)
-               {
-                    $user = User::findById($duserid);
-                    if(  $user->delete() == 1) { //1 row affect by delete, as expect..
-                      $deleted[] = $user->getId();
-                    }
-               }
-               $this->app->flash('info', 'Users with IDs  ' . implode(',',$deleted) . ' have been deleted.');
+                $userlist = $request->post('userlist'); 
+                $deleted = [];
+
+                if($userlist == NULL){
+                    $this->app->flash('info','No user to be deleted.');
+                } else {
+                     foreach( $userlist as $duserid)
+                     {
+                          $user = User::findById($duserid);
+                          if(  $user->delete() == 1) { //1 row affect by delete, as expect..
+                            $deleted[] = $user->getId();
+                          }
+                     }
+                     $this->app->flash('info', 'Users with IDs  ' . implode(',',$deleted) . ' have been deleted.');
+                }
+
+                $this->app->redirect('/admin');
+            } else {
+                $username = Auth::user()->getUserName();
+                $this->app->flash('info', 'You do not have access this resource. You are logged in as ' . $username);
+                $this->app->redirect('/');
+            }
+          }else {
+              $this->app->redirect('/');
           }
-
-          $this->app->redirect('/admin');
-      } else {
-          $username = Auth::user()->getUserName();
-          $this->app->flash('info', 'You do not have access this resource. You are logged in as ' . $username);
-          $this->app->redirect('/');
-      }
     }
-
 
     function show($tuserid)   
     {
@@ -149,43 +156,50 @@ class UserController extends Controller
 
     function edit($tuserid)    
     { 
-
-        $user = User::findById($tuserid);
-
-        if (! $user) {
-            throw new \Exception("Unable to fetch logged in user's object from db.");
-        } elseif (Auth::userAccess($tuserid)) {
-
-
-            $request = $this->app->request;
-
-            $username = $request->post('username');
-            $password = $request->post('password');
-            $email = $request->post('email');
-            $bio = $request->post('bio');
-
-            $isAdmin = ($request->post('isAdmin') != null);
-            
-
-            $user->setUsername($username);
-            $user->setPassword($password);
-            $user->setBio($bio);
-            $user->setEmail($email);
-            $user->setIsAdmin($isAdmin);
-
-            $user->save();
-            $this->app->flashNow('info', 'Your profile was successfully saved.');
+        $request = $this->app->request;
+        $usrStr = $request->post('str');
+        if($usrStr == $_SESSION["randStr"]) {
+                
 
             $user = User::findById($tuserid);
 
-            $this->render('showuser.twig', ['user' => $user]);
+            if (! $user) {
+                throw new \Exception("Unable to fetch logged in user's object from db.");
+            } elseif (Auth::userAccess($tuserid)) {
 
 
-        } else {
-            $username = $user->getUserName();
-            $this->app->flash('info', 'You do not have access this resource. You are logged in as ' . $username);
+
+                $username = $request->post('username');
+                $password = $request->post('password');
+                $email = $request->post('email');
+                $bio = $request->post('bio');
+
+                $isAdmin = ($request->post('isAdmin') != null);
+
+
+                $user->setUsername($username);
+                $user->setPassword($password);
+                $user->setBio($bio);
+                $user->setEmail($email);
+                $user->setIsAdmin($isAdmin);
+
+                $user->save();
+                $this->app->flashNow('info', 'Your profile was successfully saved.');
+
+                $user = User::findById($tuserid);
+
+                $this->render('showuser.twig', ['user' => $user]);
+
+
+            } else {
+                $username = $user->getUserName();
+                $this->app->flash('info', 'You do not have access this resource. You are logged in as ' . $username);
+                $this->app->redirect('/');
+            }
+        }else {
             $this->app->redirect('/');
         }
+        
     }
 
 }
